requirements.txt
paho-mqtt==1.6.1
pygame==2.5.2
pydub==0.25.1
gTTS==2.5.1